refactor: Service 예약, 시간 분리
