+++ src/main/java/roomescape/service/ReservationService.java
@@ -1,7 +1,6 @@
 package roomescape.service;
 
 import java.time.LocalDate;
-import java.time.LocalTime;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import roomescape.domain.Reservation;
@@ -37,22 +36,5 @@ public class ReservationService {
 
     public void deleteReservation(Long id) {
         reservationDAO.deleteById(id);
-    }
-
-    public List<ReservationTime> getReservationTimes() {
-        return reservationTimeDAO.findAll();
-    }
-
-    public ReservationTime createReservationTime(LocalTime startAt) {
-        ReservationTime reservationTime = new ReservationTime(
-                null,
-                startAt
-        );
-        Long newReservationTimeId = reservationTimeDAO.save(reservationTime);
-        return reservationTimeDAO.findById(newReservationTimeId);
-    }
-    
-    public void deleteReservationTime(Long id) {
-        reservationTimeDAO.deleteById(id);
     }
 }
